+++ resources/libraries/helper.rb
@@ -42,23 +42,14 @@ module Firewall
           permanent true
           notifies :reload, 'service[firewalld]', :delayed
         end
-      when :filter_by_ip
+      when :network
         act = value[:action]
-        name = value[:name]
-        port_val = value[:port]
-        ip_val = value[:ip]
-
-        unless valid_ip?(ip_val)
-          Chef::Log.warn('Firewall rule will not be applied.')
-          return
-        end
-
-        rich_rule = "rule family='ipv4' source address='#{ip_val}' port port='#{port_val}' protocol='#{protocol}' accept"
-        firewall_rule "#{act} #{name} port #{port_val}/#{protocol} for IP: #{ip_val} in #{zone} zone" do
-          rules rich_rule
+        firewall_rule "#{act} source #{value[:network]} in #{zone} zone" do
+          sources value[:network]
           zone zone
           action act
           permanent true
+          notifies :reload, 'service[firewalld]', :delayed
         end
       end
     end
@@ -99,7 +90,7 @@ module Firewall
     end
 
     def get_existing_rules_in_zone(zone)
-      rich_rules = shell_out!("firewall-cmd --zone=#{zone} --list-rich-rules").stdout
+      rich_rules = shell_out!("firewall-cmd --permanent --zone=#{zone} --list-rich-rules").stdout
       existing_rules = []
       rich_rules.split("\n").each do |rule|
         existing_rules << rule
@@ -190,26 +181,55 @@ module Firewall
       allowed_ips.uniq.compact
     end
 
+    # Returns a list of IPs of vault sensors that are allowed to send syslog to the proxy.
+    # vault_sensor_in_proxy_nodes comes from each proxy role's sensors_mapping['vault'],
+    # so every element is a Hash like { hostname => redborder_attrs }.
     def get_ips_allowed_for_syslog_in_proxy(vault_sensor_in_proxy_nodes)
       allowed_ips = []
       proxy_id = node['redborder']['sensor_id']
 
-      (vault_sensor_in_proxy_nodes || []).each do |sensor_node|
-        sensor_info = sensor_node.to_hash
+      (vault_sensor_in_proxy_nodes || []).each do |sensor_info|
         next unless sensor_info.is_a?(Hash)
 
-        parent_id = sensor_info['redborder']['parent_id']
-        ip = sensor_info['ipaddress']
+        sensor_info.each do |_hostname, data|
+          next unless data.is_a?(Hash)
 
-        # Just add the IP if it matches the parent_id and is a valid IPv4 address
-        if parent_id.to_i == proxy_id.to_i && ip =~ /^\d{1,3}(\.\d{1,3}){3}$/
-          allowed_ips << ip
-        else
-          Chef::Log.warn(">> [Proxy] Sensor omitted: IP=#{ip.inspect}, parent_id=#{parent_id}")
+          parent_id = data['parent_id']
+          ip = data['ipaddress']
+
+          # Just add the IP if it matches the parent_id and is a valid IPv4 address
+          if parent_id.to_i == proxy_id.to_i && ip =~ /^\d{1,3}(\.\d{1,3}){3}$/
+            allowed_ips << ip
+          else
+            Chef::Log.warn(">> [Syslog Proxy] Sensor omitted: IP=#{ip.inspect}, parent_id=#{parent_id}")
+          end
         end
       end
 
       allowed_ips.uniq.compact
+    end
+
+    # Returns the IPs of all vault sensors that report to ANY proxy. Used by the
+    # manager to avoid opening port 514 for vaults already handled by a proxy.
+    # Authoritative: comes from the proxies' sensors_mapping, so it does not depend
+    # on each vault sensor having run chef-client after being moved.
+    def get_vault_ips_in_proxies(vault_sensor_in_proxy_nodes)
+      vault_sensor_in_proxy_nodes.flat_map { |h| h.values.map { |v| v['ipaddress'] } }.compact.uniq
+    end
+
+    def get_existing_sources(zone)
+      networks = shell_out!("firewall-cmd --permanent --zone=#{zone} --list-sources").stdout
+      existing_networks = []
+      networks.split(' ').each do |network|
+        existing_networks << network
+      end
+      existing_networks
+    end
+
+    def zone_exists?(zone)
+      shell_out!('firewall-cmd --get-zones').stdout.split.include?(zone)
+    rescue Mixlib::ShellOut::ShellCommandFailed
+      false
     end
   end
 end